third_party/blink/web_tests/external/wpt/css/css-align/blocks/align-content-block-006-expected.txt

This is a testharness.js-based test.
[PASS] .test 1: start
[FAIL] .test 2: center
  assert_equals: \n<div class="test" style="align-content: center" title="center">\n    <div class="float" data-offset-y="15">FLT</div>\n    <div class="in-flow" data-offset-y="25"></div>\n    <div class="in-flow">\n      <div class="float" data-offset-y="25">FLT</div>\n      <span class="label">CENTER</span>\n      <span class="abspos">ABS</span>\n      <span class="relpos">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 15 but got 5
[FAIL] .test 3: end
  assert_equals: \n<div class="test" style="align-content: end" title="end">\n    <div class="float" data-offset-y="25">FLT</div>\n    <div class="in-flow" data-offset-y="35"></div>\n    <div class="in-flow">\n      <div class="float" data-offset-y="35">FLT</div>\n      <span class="label">END</span>\n      <span class="abspos">ABS</span>\n      <span class="relpos">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 25 but got 5
[PASS] .test 4: baseline
[FAIL] .test 5: last baseline
  assert_equals: \n<div class="test" style="align-content: last baseline" title="last baseline">\n    <div class="float" data-offset-y="25">FLT</div>\n    <div class="in-flow" data-offset-y="35"></div>\n    <div class="in-flow">\n      <div class="float" data-offset-y="35">FLT</div>\n      <span class="label">LAST BASELINE</span>\n      <span class="abspos">ABS</span>\n      <span class="relpos">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 25 but got 10
[PASS] .test 6: flex-start
[FAIL] .test 7: flex-end
  assert_equals: \n<div class="test" style="align-content: flex-end" title="flex-end">\n    <div class="float" data-offset-y="25">FLT</div>\n    <div class="in-flow" data-offset-y="35"></div>\n    <div class="in-flow">\n      <div class="float" data-offset-y="35">FLT</div>\n      <span class="label">FLEX-END</span>\n      <span class="abspos">ABS</span>\n      <span class="relpos">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 25 but got 5
[PASS] .test 8: unsafe start
[FAIL] .test 9: unsafe center
  assert_equals: \n<div class="test" style="align-content: unsafe center" title="unsafe center">\n    <div class="float" data-offset-y="15">FLT</div>\n    <div class="in-flow" data-offset-y="25"></div>\n    <div class="in-flow">\n      <div class="float" data-offset-y="25">FLT</div>\n      <span class="label">UNSAFE CENTER</span>\n      <span class="abspos">ABS</span>\n      <span class="relpos">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 15 but got 5
[FAIL] .test 10: unsafe end
  assert_equals: \n<div class="test" style="align-content: unsafe end" title="unsafe end">\n    <div class="float" data-offset-y="25">FLT</div>\n    <div class="in-flow" data-offset-y="35"></div>\n    <div class="in-flow">\n      <div class="float" data-offset-y="35">FLT</div>\n      <span class="label">UNSAFE END</span>\n      <span class="abspos">ABS</span>\n      <span class="relpos">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 25 but got 5
[PASS] .test 11: safe start
[FAIL] .test 12: safe center
  assert_equals: \n<div class="test" style="align-content: safe center" title="safe center">\n    <div class="float" "="" data-offset-y="15">FLT</div>\n    <div class="in-flow" "="" data-offset-y="25"></div>\n    <div class="in-flow">\n      <div class="float" "="" data-offset-y="25">FLT</div>\n      <span class="label">SAFE CENTER</span>\n      <span class="abspos">ABS</span>\n      <span class="relpos">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 15 but got 5
[FAIL] .test 13: safe end
  assert_equals: \n<div class="test" style="align-content: safe end" title="safe end">\n    <div class="float" data-offset-y="25">FLT</div>\n    <div class="in-flow" data-offset-y="35"></div>\n    <div class="in-flow">\n      <div class="float" data-offset-y="35">FLT</div>\n      <span class="label">SAFE END</span>\n      <span class="abspos">ABS</span>\n      <span class="relpos">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 25 but got 5
[FAIL] .test 14: space-evenly
  assert_equals: \n<div class="test" style="align-content: space-evenly" title="space-evenly">\n    <div class="float" data-offset-y="15">FLT</div>\n    <div class="in-flow" data-offset-y="25"></div>\n    <div class="in-flow">\n      <div class="float" data-offset-y="25">FLT</div>\n      <span class="label">SPACE-EVENLY</span>\n      <span class="abspos">ABS</span>\n      <span class="relpos">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 15 but got 5
[PASS] .test 15: space-between
[FAIL] .test 16: space-around
  assert_equals: \n<div class="test" style="align-content: space-around" title="space-around">\n    <div class="float" data-offset-y="15">FLT</div>\n    <div class="in-flow" data-offset-y="25"></div>\n    <div class="in-flow">\n      <div class="float" data-offset-y="25">FLT</div>\n      <span class="label">SPACE-AROUND</span>\n      <span class="abspos">ABS</span>\n      <span class="relpos">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 15 but got 5
[PASS] .test 17: normal
Harness: the test ran to completion.
